% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{CC}
\alias{CC}
\title{Function for executing CloudCompare plug-in, used in get_SEG(), taken from:
https://github.com/GreKro/cloudcompare.}
\usage{
CC(cc_function, cc_dir)
}
\arguments{
\item{cc_function}{function, cc_TREEiso()}

\item{cc_dir}{string, path to CloudCompare.exe}
}
\value{
-
}
\description{
Function for executing CloudCompare plug-in, used in get_SEG(), taken from:
https://github.com/GreKro/cloudcompare.
}
