% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{cc_TREEiso}
\alias{cc_TREEiso}
\title{Function for grabbing treeiso plug-in employed in CloudCompare, used in CC(), taken from:
https://github.com/GreKro/cloudcompare and modified.}
\usage{
cc_TREEiso(
  LAS_char,
  K1 = 5,
  L1 = 1,
  DEC_R1 = 0.05,
  K2 = 20,
  L2 = 20,
  MAX_GAP = 2,
  DEC_R2 = 0.1,
  VER_O_W = 0.5,
  RHO = 0.5,
  output_dir,
  global_shift = F,
  global_shift_type = "AUTO",
  filter_sf = F,
  filter_value = c(0, 1),
  c_export_fmt = "LAS",
  c_ext = "las",
  silent = T,
  no_timestamp = T
)
}
\arguments{
\item{LAS_char}{string, path to las file including the name of las file}

\item{output_dir}{string, path to output directory}

\item{global_shift...no_timestamp}{CloudCompare parameters: https://www.cloudcompare.org/doc/wiki/index.php/Command_line_mode}
}
\value{
las file
}
\description{
Function for grabbing treeiso plug-in employed in CloudCompare, used in CC(), taken from:
https://github.com/GreKro/cloudcompare and modified.
}
