% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{check_DAT}
\alias{check_DAT}
\title{Function for checking input data used in get_VOXEL().}
\usage{
check_DAT(data, message)
}
\arguments{
\item{data}{data.table, tibble, data.frame  as las@data}

\item{message}{logical, if TRUE error and interactive messages are enabled}
}
\value{
data table
}
\description{
Function for checking input data used in get_VOXEL().
}
