% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_3DTREE}
\alias{get_3DTREE}
\title{Function for extraction point clouds to individual tree segments.}
\usage{
get_3DTREE(lasFILE, multiPOLY, FEATURE)
}
\arguments{
\item{lasFILE}{las file of forest}

\item{multiPOLY}{sf multipolygon, individual tree segments}

\item{FEATURE}{character, attribute name to extract from multiPOLY}
}
\value{
list of las files (point clouds of individual tree segments)
}
\description{
Function for extraction point clouds to individual tree segments.
}
