% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_CBH}
\alias{get_CBH}
\title{MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.}
\usage{
get_CBH(
  list_LAS_char,
  min_RANGE = 5,
  min_POINT = 0.2,
  min_H_scale = 0.13,
  branch_WIDTH = 0.2,
  cross_WIDTH = 5,
  cbh_ONLY = 3,
  kM = TRUE,
  cbh_BUFF = 0.5,
  method = NULL,
  outdir1 = NULL,
  outdir2 = NULL,
  K1 = 10,
  L1 = 1,
  DEC_R1 = 0.1,
  K2 = 20,
  L2 = 20,
  MAX_GAP = 0.5,
  DEC_R2 = 0.1,
  VER_O_W = 0.3,
  RHO = 0.5,
  cc_dir = NULL,
  VOL = FALSE
)
}
\arguments{
\item{list_LAS_char}{character, list of input tree point cloud las files (outputted by get_3DTREE())}

\item{min_RANGE}{numeric, minimum height range (m, default = 5) of 3D tree segment employed during the process of within-segment tree isolation}

\item{min_POINT}{numeric, minimum height of points to eliminate forest floor and low vegetation (default = 0.2 m)}

\item{min_H_scale}{numeric, height scaler (m, default = .13), controlling understory removal}

\item{branch_WIDTH}{numeric, assumed CBH branch width (m, default = 0.2), controlling bin width for counting points}

\item{cross_WIDTH}{numeric, width of cross-section (m, default = 5)}

\item{cbh_ONLY}{numeric, options for executing: 1~treeiso and cbh, 2~only treeiso, 3~only cbh detection (default = 3, meaning CBH detection is active)}

\item{kM}{logical, automatic percentile-based (default 'kM' = TRUE) or interactive CBH tuning}

\item{cbh_BUFF}{logical, if 'kM' = FALSE, the vertical buffer around the CBH (CBH-cbh_BUFF and CBH+cbh_BUFF) displayed with dotted red lines (default = 0.5)}

\item{method}{character, optional additional attribute (default = NULL in combination with default 'cbh_ONLY = 3')}

\item{outdir1}{string, path to output directory of treeiso segment results (default = NULL in combination with default 'cbh_ONLY = 3')}

\item{outdir2}{string, path to output directory of filtered segments (intermediate_segs and final_segs) (default = NULL)}

\item{K1, L1, DEC_R1}{first stage cut-pursuit parameters (treeiso), default values as indicated}

\item{K2, L2, MAX_GAP, DEC_R2}{second stage cut-pursuit parameters (treeiso), default values as indicated}

\item{VER_O_W, RHO}{final stage treeiso parameters, default values as indicated}

\item{cc_dir}{string, path to CloudCompare.exe (default = NULL in combination with default 'cbh_ONLY = 3')}

\item{VOL}{logical, if TRUE returns Hull_area, Del_vol, Cube_vol and Sphere_vol too (default = FALSE)}
}
\value{
tibble (Z_max, Z_mean, Z_sd, Z_N_points, N_points, CBH and treeID) plus (Hull_area, Del_vol, Cube_vol, Sphere_vol)
}
\description{
MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.
}
