% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_CROSS}
\alias{get_CROSS}
\title{Function for getting 2D cross-section from 3D point cloud, used in get_CBH().}
\usage{
get_CROSS(las, cross_WIDTH = 5)
}
\arguments{
\item{las}{las file}

\item{cross_WIDTH}{numeric, width of cross-section, metric, default = 5}
}
\value{
las file
}
\description{
Function for getting 2D cross-section from 3D point cloud, used in get_CBH().
}
