% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_PARAMS}
\alias{get_PARAMS}
\title{Automated parameter configuration for get_CBH() based on point cloud density}
\usage{
get_PARAMS(input_dir, density_threshold = 20)
}
\arguments{
\item{input_dir}{string, path to directory containing LAS files}

\item{density_threshold}{numeric, point density threshold (points/m²) for automatic vs interactive mode (default = 20)}
}
\value{
data.frame with recommended parameters for get_CBH
}
\description{
Automated parameter configuration for get_CBH() based on point cloud density
}
