% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_SEG}
\alias{get_SEG}
\title{Function for implementing within segment tree isolation using treeiso, used in get_CBH():
Xi, Z.; Hopkinson, C. 3D Graph-Based Individual-Tree Isolation (Treeiso) from Terrestrial Laser Scanning Point Clouds.
Remote Sens. 2022, 14, 6116. https://doi.org/10.3390/rs14236116}
\usage{
get_SEG(
  list_LAS_char,
  outdir1,
  outdir2,
  min_RANGE = 5,
  min_POINT = 0.2,
  K1 = 10,
  L1 = 1,
  DEC_R1 = 0.1,
  K2 = 20,
  L2 = 20,
  MAX_GAP = 0.5,
  DEC_R2 = 0.1,
  VER_O_W = 0.3,
  RHO = 0.5,
  cc_dir
)
}
\arguments{
\item{list_LAS_char}{character, list of las files}

\item{outdir1}{string, path to output directory of tree isolation (treeiso) segments}

\item{outdir2}{string, path to output directory of filtered segments (intermediate_segs and final_segs)}

\item{min_RANGE}{numeric, minimum height range (m) of 3D tree segment employed during the process of within-segment tree isolation, default = 5}

\item{min_POINT}{numeric, minimum height of points to eliminate forest floor and low vegetation (default = 0.2 m)}

\item{K1, L1, DEC_R1, K2, L2, MAX_GAP, DEC_R2, VER_O_W, RHO}{see function get_CBH()}

\item{cc_dir}{string, path to CloudCompare.exe}
}
\value{
las file

las file
}
\description{
Function for implementing within segment tree isolation using treeiso, used in get_CBH():
Xi, Z.; Hopkinson, C. 3D Graph-Based Individual-Tree Isolation (Treeiso) from Terrestrial Laser Scanning Point Clouds.
Remote Sens. 2022, 14, 6116. https://doi.org/10.3390/rs14236116
}
