% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_VOXEL}
\alias{get_VOXEL}
\title{Helper functions from VOXR package:
Bastien Lecigne and others, Exploring trees in three dimensions:
VoxR, a novel voxel-based R package dedicated to analysing the complex arrangement of tree crowns,
Annals of Botany, Volume 121, Issue 4, 14 March 2018, Pages 589–601,
https://doi.org/10.1093/aob/mcx095}
\usage{
get_VOXEL(data, res_VOXEL, full.grid, message)
}
\arguments{
\item{data}{data.table, tibble, data.frame  as las@data}

\item{res_VOXEL}{numeric, voxel resolution}

\item{full.grid}{logical, if TRUE empty voxels in the tree bounding box are returned}

\item{message}{logical, if TRUE error and interactive messages are enabled}
}
\value{
data table
}
\description{
Function for voxelizing point cloud.
}
