% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{plot_CROSS}
\alias{plot_CROSS}
\title{Function for 2D cross-sectional plot used only in get_CBH().}
\usage{
plot_CROSS(las, col = "grey25", ylab, cross_WIDTH = 5)
}
\arguments{
\item{las}{las file}

\item{col}{character, color of points, default = "grey25"}

\item{cross_WIDTH}{numeric, width of cross-section (m, default = 5)}
}
\value{
displays ggplot
}
\description{
Function for 2D cross-sectional plot used only in get_CBH().
}
