% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_3DTREE}
\alias{get_3DTREE}
\title{Function for extraction point clouds to individual tree segments.}
\usage{
get_3DTREE(
  lasFILE,
  multiPOLY,
  normalize = TRUE,
  output_dir,
  FEATURE = NULL,
  RETURN = FALSE
)
}
\arguments{
\item{lasFILE}{las file of forest}

\item{multiPOLY}{sf multipolygon, individual tree segments}

\item{normalize}{logical, if TRUE normalization is performed on the ENTIRE dataset before segmentation}

\item{output_dir}{string, path to output directory}

\item{FEATURE}{character, attribute name and its values}

\item{RETURN}{logical, whether to return the list of las or not (default = FALSE)}
}
\value{
list of las files (point clouds of individual tree segments)
}
\description{
Function for extraction point clouds to individual tree segments.
}
