% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_3DTREE}
\alias{get_3DTREE}
\title{Function for extraction point clouds to individual tree segments.}
\usage{
get_3DTREE(lasFILE, multiPOLY, normalize = TRUE, output_dir, FEATURE)
}
\arguments{
\item{lasFILE}{las file of forest}

\item{multiPOLY}{sf multipolygon, individual tree segments}

\item{normalize}{logical, normalization based on CSF-classified ground points using the lidR::knnidw(), if FALSE "flat normalization" is performed}

\item{output_dir}{string, path to output directory}

\item{FEATURE}{character, attribute name to extract from multiPOLY to store in las}
}
\value{
list of las files (point clouds of individual tree segments)
}
\description{
Function for extraction point clouds to individual tree segments.
}
