% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_CBH}
\alias{get_CBH}
\title{MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.}
\usage{
get_CBH(
  list_LAS_char,
  min_RANGE = 5,
  min_H_scale = 0.13,
  branch_WIDTH = 0.5,
  cross_WIDTH = 5,
  ONLY = FALSE,
  kM = FALSE,
  method = NULL,
  outdir1,
  outdir2,
  K1 = 10,
  L1 = 1,
  DEC_R1 = 0.1,
  K2 = 20,
  L2 = 20,
  MAX_GAP = 0.5,
  DEC_R2 = 0.1,
  VER_O_W = 0.3,
  RHO = 0.5,
  ret = F
)
}
\arguments{
\item{etc.}{see function}
}
\value{
tibble (Z_max, Z_mean, Z_sd, Z_N_points, N_points, CBH, Hull_area, Del_vol, Cube_vol, Sphere_vol and treeID)
}
\description{
MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.
}
