% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecbh.R
\name{get_CBH}
\alias{get_CBH}
\title{MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.}
\usage{
get_CBH(
  list_LAS_char,
  min_RANGE = 5,
  min_H_scale = 0.13,
  branch_WIDTH = 0.2,
  cross_WIDTH = 5,
  ONLY = FALSE,
  kM = FALSE,
  method = NULL,
  outdir1,
  outdir2,
  K1 = 10,
  L1 = 1,
  DEC_R1 = 0.1,
  K2 = 20,
  L2 = 20,
  MAX_GAP = 0.5,
  DEC_R2 = 0.1,
  VER_O_W = 0.3,
  RHO = 0.5,
  cc_dir
)
}
\arguments{
\item{list_LAS_char}{character, list of las files}

\item{min_RANGE}{numeric, minimum height range (m, default = 5) of 3D tree segment employed during the process of within-segment tree isolation}

\item{min_H_scale}{numeric, height scaler (m, default = .13), controlling understory removal}

\item{branch_WIDTH}{numeric, assumed CBH branch width (m, default = 0.2), controlling bin width for counting points}

\item{cross_WIDTH}{numeric, width of cross-section (m, default = 5)}

\item{ONLY}{logical, whether disable treeiso or not (default = FALSE, meaning tree isolation is enabled)}

\item{kM}{logical, interactice K-means cluster k tuning, activated if 'kM' = TRUE (default = FALSE)}

\item{method}{character, optional additional attribute (default = NULL)}

\item{outdir1}{string, path to output directory of treeiso segment results}

\item{outdir2}{string, path to output directory of filtered segments (intermediate_segs and final_segs)}

\item{K1, L1, DEC_R1}{first stage cut-pursuit parameters (treeiso), default values as indicated}

\item{VER_O_W, RHO}{final stage treeiso parameters, default values as indicated}

\item{cc_dir}{string, path to CloudCompare.exe}

\item{K2-L2, MAX_GAP, DEC_R2}{second stage cut-pursuit parameters (treeiso), default values as indicated}
}
\value{
tibble (Z_max, Z_mean, Z_sd, Z_N_points, N_points, CBH, Hull_area, Del_vol, Cube_vol, Sphere_vol and treeID)
}
\description{
MAIN FUNCTION of treecbh, detecting CBH and deriving numerous metrics.
}
